requirements.txt
scrapy
scrapy-djangoitem
scrapyd
ipdb
ipython
celery
django
django-celery
djangorestframework
